MOR-452: Re-organised component files
