+++ src/components/editor/BarItemFields.vue
