+++ src/components/editor/BarItemLink.vue
