+++ src/components/editor/ButtonCatalog.vue
