+++ src/components/side-panel/BarsList.vue
